requirements.txt
aiohttp
Pillow
beautifulsoup4
lxml